service/requirements.txt
async-lru==2.0.5
fastapi[standard]==0.115.12
google-auth==2.40.3
psycopg2-binary==2.9.10
pyutils @ git+https://github.com/iagocanalejas/pyutils.git@master
requests==2.32.4
